Updates signature on new SDW Buster template
The first Buster-based SDW template image was signed with the FPF
Authority Key, rather than the SecureDrop TEST signing key. We use the
FPF Authority Key to sign the git tags for the template builds, but the
RPMs themselves are signed with the SD test key for now. (We'll likely
transition to the prod SD release signing key in the near future.)