Removes older 2019 template
Preserving 2 recent template RPMs, for a modicum of rollback capability.
Practically speaking, the older template that's preserved here isn't
functional, given that it contains an expired apt pubkey, but we'll soon
by updating the template again with new kernels, so the pragma of 2
template RPMs is useful to enforce now.