Strip signatures from RPMs
RPMs in this repository are no longer supposed to be signed but
these pre-date that. But having them signed means that different
RPM tooling will try to validate the signature, which fails if
you don't have the test repository key imported and it's just
easier to remove the remaining signatures and assume the packages
will be unsigned going forwards.