Add new repo structure based on Qubes releases
Adds symbolic links, linking the previous repository structure to the
new one. This will enable systems to do an inplace upgrade
from Qubes 4.2 to 4.3 by using a '$releasever' substitution in its
'.repo' configuration. See discussion #1508

By keeping the old directory structure, systems will still be able to
fetch packages in the old format. This is particularly relevant for
fetching the new keyring repo, to be released with this change [1],
which gives existing systems information about the new repository
structure.

[1]: https://github.com/freedomofpress/securedrop-workstation-keyring/pull/38